[CodePartition] Move buffer related logic to WSBuffer (#101)
Also add WSUtility.h, WSUtility.cpp
